refactor(clp-s): Split clp-s into several smaller libraries to ease re-use in other projects. (#924)
