+++ components/core/src/clp_s/JsonConstructor.cpp
@@ -10,11 +10,10 @@
 #include <mongocxx/exception/exception.hpp>
 #include <mongocxx/uri.hpp>
 #include <nlohmann/json.hpp>
+#include <spdlog/spdlog.h>
 
 #include "archive_constants.hpp"
 #include "ErrorCode.hpp"
-#include "ReaderUtils.hpp"
-#include "SchemaTree.hpp"
 #include "TraceableException.hpp"
 
 namespace clp_s {
