+++ components/core/src/clp_s/VariableDecoder.cpp
@@ -2,6 +2,8 @@
 
 #include "VariableDecoder.hpp"
 
+#include <spdlog/spdlog.h>
+
 namespace clp_s {
 bool VariableDecoder::decode_variables_into_message(
         LogTypeDictionaryEntry const& logtype_dict_entry,
