+++ components/core/src/clp_s/ZstdCompressor.hpp
@@ -6,7 +6,6 @@
 #include <memory>
 #include <string>
 
-#include <spdlog/spdlog.h>
 #include <zstd.h>
 #include <zstd_errors.h>
