+++ components/core/src/clp_s/search/Output.cpp
@@ -3,6 +3,8 @@
 #include <memory>
 #include <vector>
 
+#include <spdlog/spdlog.h>
+
 #include "../../clp/type_utils.hpp"
 #include "../SchemaTree.hpp"
 #include "../Utils.hpp"
