+++ components/core/src/clp_s/search/Output.hpp
@@ -9,8 +9,6 @@
 #include <unordered_set>
 #include <utility>
 
-#include <simdjson.h>
-
 #include "../ArchiveReader.hpp"
 #include "../SchemaReader.hpp"
 #include "../Utils.hpp"
@@ -21,7 +19,6 @@
 #include "QueryRunner.hpp"
 #include "SchemaMatch.hpp"
 
-using namespace simdjson;
 using namespace clp_s::search::clp_search;
 
 namespace clp_s::search {
