+++ components/core/src/clp_s/search/ast/CMakeLists.txt
@@ -56,10 +56,11 @@ add_library(
 )
 add_library(clp_s::search::ast ALIAS clp_s_search_ast)
 target_compile_features(clp_s_search_ast PRIVATE cxx_std_20)
+target_include_directories(clp_s_search_ast PUBLIC ../../../)
 target_link_libraries(
     clp_s_search_ast
     PUBLIC
-    clp_s::TimestampPattern
+    clp_s::timestamp_pattern
     PRIVATE
     simdjson::simdjson
 )
