+++ components/core/src/clp_s/search/sql/CMakeLists.txt
@@ -6,16 +6,17 @@ ANTLR_TARGET(
 )
 
 add_library(
-        sql
+        clp_s_search_sql
         ../antlr_common/ErrorListener.hpp
         ${ANTLR_SqlParser_CXX_OUTPUTS}
         sql.cpp
         sql.hpp
 )
-target_compile_features(sql PRIVATE cxx_std_20)
-target_include_directories(sql PRIVATE ${ANTLR_SqlParser_OUTPUT_DIR})
+add_library(clp_s::search::sql ALIAS clp_s_search_sql)
+target_compile_features(clp_s_search_sql PRIVATE cxx_std_20)
+target_include_directories(clp_s_search_sql PRIVATE ${ANTLR_SqlParser_OUTPUT_DIR})
 target_link_libraries(
-	sql
+	clp_s_search_sql
 	PRIVATE
 	antlr4_static
         clp_s::search::ast
