+++ components/core/tests/clp_s_test_utils.cpp
@@ -5,7 +5,6 @@
 
 #include <catch2/catch.hpp>
 
-#include "../src/clp_s/CommandLineArguments.hpp"
 #include "../src/clp_s/InputConfig.hpp"
 #include "../src/clp_s/JsonParser.hpp"
 
@@ -14,7 +13,7 @@ void compress_archive(
         std::string const& archive_directory,
         bool single_file_archive,
         bool structurize_arrays,
-        clp_s::CommandLineArguments::FileType file_type
+        clp_s::FileType file_type
 ) {
     constexpr auto cDefaultTargetEncodedSize{8ULL * 1024 * 1024 * 1024};  // 8 GiB
     constexpr auto cDefaultMaxDocumentSize{512ULL * 1024 * 1024};  // 512 MiB
@@ -40,9 +39,9 @@ void compress_archive(
     parser_option.input_file_type = file_type;
 
     clp_s::JsonParser parser{parser_option};
-    if (clp_s::CommandLineArguments::FileType::Json == file_type) {
+    if (clp_s::FileType::Json == file_type) {
         REQUIRE(parser.parse());
-    } else if (clp_s::CommandLineArguments::FileType::KeyValueIr == file_type) {
+    } else if (clp_s::FileType::KeyValueIr == file_type) {
         REQUIRE(parser.parse_from_ir());
     } else {
         // This branch should be unreachable.
