+++ taskfiles/lint.yaml
@@ -429,6 +429,8 @@ tasks:
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/JsonParser.cpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/JsonParser.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/JsonSerializer.hpp"
+            - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/OutputHandlerImpl.cpp"
+            - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/OutputHandlerImpl.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/PackedStreamReader.cpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/PackedStreamReader.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/ParsedMessage.hpp"
@@ -492,8 +494,6 @@ tasks:
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/kql/kql.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/Output.cpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/Output.hpp"
-            - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/OutputHandler.cpp"
-            - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/OutputHandler.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/Projection.cpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/Projection.hpp"
             - "{{.G_CORE_COMPONENT_DIR}}/src/clp_s/search/QueryRunner.cpp"
